Add verity test for when using ignore_index
Should default the proper output to zero but give the same result. Updated gold data and unit tests to handle this too.

Using the final mul and default values to act as a mask instead of using where since we always know weights are 1d thus, just mask off the one weight for a given class label at the proper index with a scatter_none