Add additional verify tests and fix gen of onnx files + comments
-Add additional test cases when batch > class_size and class_size < batch - These currently fail and need to be handled
-Fix k-dimensional reduction output for tests and regnerate these
-Add comment in parser in how we select the weight output

TODO fix output scatter. it appears for asymmetric case the last input never gets populated correctly. May need to pad the input or figure out a better way instead of just doing a scatter. seeems like there's a clever way to handle things here if say, class size = batch size we can reduce things to a simple scatter_none op, but in the asymmetric case we run into issues with the output.

I've added the gold data scratchpad updated for these two assymmetric cases as well and used the output to populate the test data based on what the onnx ref is giving.